Implement unit tests support for OpenCL implementation.
This adds a interface for the OpenCL implementation very similar to the one that exists for SW/OPTSW/HW, and uses this to run the tests.
Care needs to be taken about being able to access this interface from C, also about the fact that the OpenCL implementation doesn't have a compressor.
This also takes care to add support for passing in/out the ilen/olen/returnValues parameters to the OpenCL implementation, just for the test's purposes.
All tests pass except the CRC one, because the OpenCL implementation doesn't have a CRC implementation.