Update OpenCL C++ bindings: cl.hpp -> cl2.hpp.
The cl.hpp header from the C++ binding is deprecated and it's quite trivial to update to cl2.hpp. cl2.hpp is also available at least on Ubuntu 16.04+ and Debian stretch+ so it should be everywhere now.

Changes for the upgrade are quite trivial:
* Change include of CL/cl.hpp -> CL/cl2.hpp
* Rename __CL_ENABLE_EXCEPTIONS -> CL_HPP_ENABLE_EXCEPTIONS
* Rename VECTOR_CLASS -> cl::vector

Additionally, for the CI builds, on Ubuntu 16.04 cl.hpp is not available, but cl2.hpp is, so that's one problem less to care about.