Fix build on old/ancient versions of the supported toolchains: * Do not rely on the default C standard of GCC (which is not C11 on GCC4). * Avoid using capture-and-assign on OpenMP (only available on 4.0+).
Build verified on Ubuntu 12.04 + GCC 4.8 + Intel OpenCL 1.2 + latest cl2.hpp as of 20200510.

NB: In reality, there's a missing error: Support for aligned_alloc is missing on Ubuntu 12.04's glibc-2.15, it's only supported on glibc-2.16. But since that's not our fault (rather missing support of the C11 standard on old glibc), it's not really worth to work around it.