src/ocl: Use OpenCL profiling to time the command queue commands
This is practically equivalent to the current timing code (timing outside OpenCL) in most cases, but is more accurate, since it will only time the time the decompression is actually executing and not other commands, processes or delays. Generally in case the command queue is empty this should be immediately and make no difference, but for example, in pocl, a cold uncached run takes ~500ms extra even with an empty command queue. This avoids the problem. Also move the timing logic to CLHostDecompressor (since it's only used there).