Introduce alignment requirements
This introduces two values which specify the required (precondition for calling) alignment and preferred (for best performance) alignment of buffers passed to 842 compressors and decompressors.

While this 'pollutes' most of the previously alignment-oblivious code with additional logic, this is necessary for efficient hardware (cryptodev) support, since buffers need to be page-aligned for best performance.
Other than this, there were previously already other reasons that justified this change:
* The optimized software compressor had a 'hidden' precondition of a 8-byte aligned input buffer. Otherwise it crashed on platforms not supporting unaligned access like ARMv7), or had terrible performance.
* The required alignment required by the hardware (cryptodev) implementation is determined by the kernel (cryptodev's alignmask). Currently no required alignment was required, but technically the alignment requirement should be respected and it could change in the future.
* The AIX hardware-accelerated implementation imposes requirements (128 byte alignment) which are typically not guaranteed by default allocation routines (malloc, new).