src/ocl/decompress.cl, src/serial_optimized/842_decompress.cpp: Always avoid out-of-bounds shift in read_bits
The OpenCL implementation did an out-of-bounds shift (shift by 64 bits), which is undefined by the C standard. This works so far on our target hardware, but it's better to avoid it.

The C++ implementation had this behind a flag, but having seen that recent versions of g++ can actually do weird things to abuse of out-of-bounds shift, just make it always conformant. This also optimizes it to be branch-free and hopefully more easily optimizable by the compiler.