src/serial_optimized/842_compress.cpp, bitstream.cpp: Use exceptions for detecting full bitstream (ENOSPC).
According to my tests, using exceptions is faster. Compared to no error checking, exception overhead is ~0.5%, while flag checks are ~1.5% with a file with 50% compressible and 50% uncompressible blocks.
Makes also sense since C++ exceptions are typically zero-cost if not thrown, while checking a flag every 8 bytes of input is not.